+++ packages/@uppy/react/src/index.ts
@@ -1,10 +1,9 @@
-export { default as Dashboard } from './Dashboard.js'
-export { default as DashboardModal } from './DashboardModal.js'
-export * from './headless/generated/index.js'
 // Headless components
+export * from './headless/generated/index.js'
 export {
   UppyContext,
   UppyContextProvider,
+  useUppyContext,
 } from './headless/UppyContextProvider.js'
 export { useDropzone } from './useDropzone.js'
 export { useFileInput } from './useFileInput.js'