+++ packages/@uppy/tus/src/index.ts
@@ -14,6 +14,7 @@ import {
   getAllowedMetaFields,
   hasProperty,
   isNetworkError,
+  type LocalUppyFile,
   NetworkError,
   RateLimitedQueue,
 } from '@uppy/utils'
@@ -97,7 +98,10 @@ type Opts<M extends Meta, B extends Body> = DefinePluginOpts<
 >
 
 declare module '@uppy/utils' {
-  export interface UppyFile<M extends Meta, B extends Body> {
+  export interface LocalUppyFile<M extends Meta, B extends Body> {
+    tus?: TusOpts<M, B>
+  }
+  export interface RemoteUppyFile<M extends Meta, B extends Body> {
     tus?: TusOpts<M, B>
   }
 }
@@ -204,7 +208,9 @@ export default class Tus<M extends Meta, B extends Body> extends BasePlugin<
    *    up a spot in the queue.
    *
    */
-  #uploadLocalFile(file: UppyFile<M, B>): Promise<tus.Upload | string> {
+  async #uploadLocalFile(
+    file: LocalUppyFile<M, B>,
+  ): Promise<tus.Upload | string> {
     this.resetUploaderReferences(file.id)
 
     // Create a new tus upload
@@ -439,6 +445,7 @@ export default class Tus<M extends Meta, B extends Body> extends BasePlugin<
 
       uploadOptions.metadata = meta
 
+      if (file.data == null) throw new Error('File data is empty')
       upload = new tus.Upload(file.data, uploadOptions)
       this.uploaders[file.id] = upload
       const eventManager = new EventManager(this.uppy)
@@ -539,11 +546,11 @@ export default class Tus<M extends Meta, B extends Body> extends BasePlugin<
     }
 
     return {
-      ...file.remote?.body,
+      ...('remote' in file && file.remote.body),
       endpoint: opts.endpoint,
       uploadUrl: opts.uploadUrl,
       protocol: 'tus',
-      size: file.data.size,
+      size: file.data!.size,
       headers: opts.headers,
       metadata: file.meta,
     }